filetype of the angles in the output changed to float instead of int as the angles may require decimal accuracies

+++ src/autokite/functions_autokite.py
@@ -162,7 +162,7 @@ def pixel_to_sky_angles(
         cy: int = 960,
         r_max: int = 960,
         projection: str = 'equidistant'
-) -> (int, int):
+) -> (float, float):
     """
     Converts pixel coordinates (x, y) from a fisheye image to sky angles (zenith θ and azimuth φ).
 
@@ -205,7 +205,7 @@ def pixel_to_angles_with_height(
         cy: int = 960,
         r_max: int = 960,
         projection: str = 'equidistant',
-) -> (int, int, (float, float, float)):
+) -> (float, float, (float, float, float)):
     """
     Convert a fisheye pixel coordinate into real-world coordinates at a known altitude.
